phase AC.8: relock thin-client bootstrap seam
