+++ boundaries/atm-daemon-client/daemon-bootstrap.toml
@@ -32,7 +32,7 @@ io_forbidden = [
 
 [dependencies]
 allowed_dependents = ["atm", "atm-daemon-bootstrap", "atm-graft"]
-allowed_dependencies = ["atm-storage"]
+allowed_dependencies = ["atm-core", "atm-storage"]
 forbidden_edges = [
   "atm-daemon-client -> atm-daemon",
   "atm-daemon-client -> atm-storage-rusqlite",
