+++ boundaries/atm-graft/shared-client-consumer.toml
@@ -30,6 +30,7 @@ io_forbidden = [
 allowed_dependents = []
 allowed_dependencies = ["atm-core", "atm-daemon-client"]
 forbidden_edges = [
+  "atm-graft -> atm-daemon-bootstrap",
   "atm-graft -> atm-daemon",
   "atm-graft -> atm-storage-rusqlite",
 ]
