+++ crates/atm-daemon-client/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 homepage.workspace = true
 
 [dependencies]
+atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.2" }
 atm-storage = { path = "../atm-storage", version = "1.2.2" }
 bytes = "1"
 fs2 = "0.4"
