+++ crates/atm-graft/Cargo.toml
@@ -11,7 +11,6 @@ homepage.workspace = true
 
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.2" }
-atm-daemon-bootstrap = { path = "../atm-daemon-bootstrap", version = "1.2.2" }
 atm-daemon-client = { path = "../atm-daemon-client", version = "1.2.2" }
 interprocess = "2.4.2"
 serde_json.workspace = true
