+++ docs/atm/architecture.md
@@ -138,8 +138,9 @@ Follow-up work:
   before daemon fork/exec
 - the CLI standard same-host bootstrap path must be the canonical thin-client
   convenience wrapper for endpoint resolution, daemon-binary resolution, probe,
-  and supervised auto-start; other first-party thin clients may mirror that
-  path, but they must not depend on CLI internals to do so
+  and supervised auto-start through the shared `atm-daemon-client` bootstrap
+  seam; other first-party thin clients may mirror that path, but they must not
+  depend on CLI internals to do so
 - the canonical thin-client bootstrap seam must stay free of runtime/storage
   composition ownership so convenience auto-start never forces `atm-runtime` or
   concrete backend crates into thin-client dependency graphs
